# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LocalAssistant API
  version: 0.1.0
  description: >-
    Educational API demonstrating an explicit language-provider and tool-calling
    orchestration loop with deterministic fake scenarios and an optional Ollama
    adapter.
  license:
    name: MIT
    url: https://opensource.org/license/mit
servers:
  - url: http://localhost:5100
    description: Local development
paths:
  /api/private/sessions:
    post:
      operationId: createPrivateSession
      summary: Exchange a registered private-client credential for a temporary loopback bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePrivateSessionRequest" }
      responses:
        "200":
          description: A temporary bearer token. The client must keep it in memory only.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PrivateSessionResponse" }
        "400": { description: The client ID or credential is missing. }
        "401": { description: The private-client credential is invalid or revoked. }
        "404": { description: The request was not received through loopback. }
  /api/private/admin/pairings:
    post:
      operationId: completePrivateClientPairing
      summary: Consume a local administrative pairing challenge and return a new client credential once.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompletePrivateClientPairingRequest" }
      responses:
        "200":
          description: The registered client and its one-time credential.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PrivateClientCredentialResponse" }
        "400": { description: The request is malformed or no installation owner exists. }
        "404": { description: "The challenge is invalid, expired, consumed, or the request is not loopback." }
  /api/private/admin/credential-rotations:
    post:
      operationId: rotatePrivateClientCredential
      summary: Consume an administrative challenge to rotate one private-client credential.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConsumeAdministrativeChallengeRequest" }
      responses:
        "200":
          description: The replacement credential for the requested client; existing sessions are invalidated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PrivateClientCredentialResponse" }
        "400": { description: The challenge or client ID is missing. }
        "404": { description: "The challenge is invalid, expired, consumed, targets another client, or the request is not loopback." }
  /api/private/admin/client-revocations:
    post:
      operationId: revokePrivateClient
      summary: Consume an administrative challenge to revoke one private client and its sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConsumeAdministrativeChallengeRequest" }
      responses:
        "200":
          description: The requested client and all of its sessions were revoked.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PrivateClientRevocationResponse" }
        "400": { description: The challenge or client ID is missing. }
        "404": { description: "The challenge is invalid, expired, consumed, targets another client, or the request is not loopback." }
...
components:
  securitySchemes:
    PrivateBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >-
        Temporary opaque token issued after a registered private client exchanges
        its durable credential through loopback.
  schemas:
    CreatePrivateSessionRequest:
      type: object
      required: [clientId, credential]
      properties:
        clientId: { type: string }
        credential: { type: string, writeOnly: true }
    PrivateSessionResponse:
      type: object
      required: [accessToken, expiresAtUtc]
      properties:
        accessToken: { type: string }
        expiresAtUtc: { type: string, format: date-time }
    CompletePrivateClientPairingRequest:
      type: object
      required: [challenge, displayName]
      properties:
        challenge: { type: string, writeOnly: true }
        displayName: { type: string, maxLength: 128 }
    ConsumeAdministrativeChallengeRequest:
      type: object
      required: [challenge, clientId]
      properties:
        challenge: { type: string, writeOnly: true }
        clientId:
          type: string
          description: The local client ID expected to match the administrative challenge target. It is not authorization.
    PrivateClientRevocationResponse:
      type: object
      required: [clientId]
      properties:
        clientId: { type: string }
    PrivateClientCredentialResponse:
      type: object
      required: [clientId, displayName, credential]
      properties:
        clientId: { type: string }
        displayName: { type: string }
        credential: { type: string }
...